% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-example250.R
\docType{data}
\name{example250}
\alias{example250}
\title{Example mediation data}
\format{
A data table with 250 rows and 6 variables:
\describe{
\item{id}{Numeric identifier for each observation (1:250)}
\item{x}{Random binary variable (0,1)}
\item{m}{Random binary variable (0,1)}
\item{y}{Random continuous variable (-3.04 to 3.54)}
\item{cens}{Random binary variable (0,1)}
\item{c}{Random continuous variable (-3.11 to 2.52)}
}
}
\usage{
example250
}
\description{
This data was randomly generated and based off the example data provided in the SAS macro <https://www.hsph.harvard.edu/tyler-vanderweele/tools-and-tutorials/>
}
\examples{
example250
}
\keyword{datasets}
