% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator.R
\name{mediator}
\alias{mediator}
\alias{mediator.default}
\title{Causal Mediation Analysis}
\usage{
mediator(...)

\method{mediator}{default}(
  data,
  out.model,
  med.model,
  treat,
  a = 1,
  a_star = 0,
  m = 0,
  boot_rep = 0,
  pm_ci = FALSE,
  ...
)
}
\arguments{
\item{...}{other arguments}

\item{data}{Data set to use for analysis}

\item{out.model}{A fitted model object for the outcome.
Can be of class 'glm','lm', or 'coxph'.}

\item{med.model}{A fitted model object for the mediator.
Can be of class 'glm','lm'.}

\item{treat}{A character string indicating the name of the
treatment/exposure variable used.}

\item{a}{A numeric value indicating the exposure level. Default = 1}

\item{a_star}{A numeric value indicating the compared exposure level.
Default = 0.}

\item{m}{A numeric value indicating the level of the mediator. Default = 0.}

\item{boot_rep}{A numeric value indicating the number of repetitions
to use when utalizing bootstrap to calculate confidence intervals.
When `boot_rep` = 0, the Delta method for calculating confidence
intervals is used. Default = 0.}

\item{pm_ci}{A logical indicator for calculating the CI for the proportion
mediated. Default = FALSE. Currently, the CI can only be determined using
boostrapping. If `pm_ci` = TRUE and `boot_rep` = 0 then 100 replicated
are automatically used.}
}
\value{
Tibble containing point estimates and 95 percent CI for the
  CDE, NDE, NIE and TE and the point estimate for the proportion mediated.
}
\description{
The `mediator` R function conducts mediation analysis under
the counterfactual framework assuming interation between the exposure
and mediator. Currently the function works for binary and continuous
outcomes and mediators.
}
\examples{
mediator(data = example250,
         out.model = lm(y ~ x + c + m + x*c, data = example250),
         med.model = lm(c ~ x + m, data = example250),
         treat = "x")

}
