% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediator.R
\name{mediator}
\alias{mediator}
\title{Causal Mediation Analysis}
\usage{
mediator(data, out.model, med.model, treat, mediator,
  out.reg = "logistic", med.reg = "logistic", a = 1, a_star = 0,
  m = 0, boot_rep = 0)
}
\arguments{
\item{data}{Data set to use for analysis}

\item{out.model}{A fitted model object for the outcome.
Can be of class 'glm','lm'.}

\item{med.model}{A fitted model object for the mediator.
Can be of class 'glm','lm'.}

\item{treat}{A character string indicating the name of the
treatment/exposure variable used.}

\item{mediator}{A character string indicating the name of the
mediator variable used.}

\item{out.reg}{A character string indicating the type of
regression used in the outcome model. Can be either "logisitic" or
"linear".}

\item{med.reg}{A character string indicating the type of
regression used in the mediator model. Can be either "logisitic" or
"linear".}

\item{a}{A numeric value indicating the exposure level. Default = 1}

\item{a_star}{A numeric value indicating the compared exposure level.
Default = 0.}

\item{m}{A numeric value indicating the level of the mediator. Default = 1}

\item{boot_rep}{A numeric value indicating the number of repetitions
to use when utalizing bootstrap to calculate confidence intervals.
When `boot_rep` = 0, the Delta method for calculating confidence
intervals is used. Default = 0.}
}
\value{
Data frame containing point estimates and 95 percent CI for the
  CDE, NDE, NIE and TE and the point estimate for the proportion mediated.
}
\description{
The `mediator` R function conducts mediation analysis under
the counterfactual framework assuming interation between the exposure
and mediator. Currently the function works for binary and continuous
outcomes and mediators.
}
